+++ speeduino/table3d_typedefs.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <stdint.h>
+
+// Table typedefs - for consistency
+typedef uint8_t table3d_dim_t;   // The x,y axis size & indices fit in this type
+typedef uint8_t table3d_value_t; // The type of each table value
+typedef int16_t table3d_axis_t;  // The type of each axis element
