+++ speeduino/updates.ino
@@ -10,7 +10,7 @@
  */
 #include "globals.h"
 #include "storage.h"
-#include "table_iterator.h"
+#include "table3d_iterator.h"
 #include EEPROM_LIB_H //This is defined in the board .h files
 
 void doUpdates()
